+++ apps/fluent-tester/src/FluentTester/TestComponents/Tabs/TabsTest.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { View } from 'react-native';
+import { Platform, View } from 'react-native';
 import { Tabs, TabsItem, Text, Button } from '@fluentui/react-native';
 import { stackStyle } from '../Common/styles';
 import { TABS_TESTPAGE } from './consts';
@@ -124,7 +124,7 @@ const tabsRenderSeparately: React.FunctionComponent = () => {
   };
 
   const getTabId = (key: string) => {
-    return `ShapeColorPivot_${key}`;
+    return `ShapeColorTabs_${key}`;
   };
 
   return (
@@ -178,6 +178,35 @@ const tabsSettingSelectedKey: React.FunctionComponent = () => {
   );
 };
 
+const tabsShowHideItem: React.FunctionComponent = () => {
+  const [showFirstItem, setshowFirstItem] = React.useState(true);
+
+  const toggleShowFirstItem = React.useCallback(() => {
+    setshowFirstItem(!showFirstItem);
+  }, [showFirstItem]);
+
+  return (
+    <View style={stackStyle}>
+      <Tabs label="Tabs">
+        {showFirstItem && (
+          <TabsItem headerText="Home" itemKey="home">
+            <Text>Click the button below to show/hide this tabs item.</Text>
+            <Text>The selected item will not change when the number of tabs items changes.</Text>
+            <Text>If the selected item was removed, the new first item will be selected.</Text>
+          </TabsItem>
+        )}
+        <TabsItem headerText="File" itemKey="file">
+          <Text>Tabs #2</Text>
+        </TabsItem>
+        <TabsItem headerText="Setting" itemKey="setting">
+          <Text>Tabs #3</Text>
+        </TabsItem>
+      </Tabs>
+      <Button content={`${showFirstItem ? 'Hide' : 'Show'} First Tabs Item`} onClick={toggleShowFirstItem} />
+    </View>
+  );
+};
+
 const tabsWithFlexibility: React.FunctionComponent = () => {
   const [selectedKey, setSelectedKey] = React.useState('home');
 
@@ -218,10 +247,6 @@ const tabsSections: TestSection[] = [
     component: disabledTabs,
   },
   {
-    name: 'Count and Icon',
-    component: tabsCountIcon,
-  },
-  {
     name: 'Trigger onTabsClick event',
     component: onTabsClickEvent,
   },
@@ -238,15 +263,26 @@ const tabsSections: TestSection[] = [
     component: tabsSettingSelectedKey,
   },
   {
+    name: 'Show/Hide Tabs item',
+    component: tabsShowHideItem,
+  },
+  {
     name: 'More Flexibility',
     component: tabsWithFlexibility,
   },
 ];
 
+if (Platform.OS !== 'windows') {
+  tabsSections.push({
+    name: 'Count and Icon',
+    component: tabsCountIcon,
+  });
+}
+
 export const TabsTest: React.FunctionComponent = () => {
   const status: PlatformStatus = {
     win32Status: 'Experimental',
-    uwpStatus: 'Backlog',
+    uwpStatus: 'Experimental',
     iosStatus: 'Backlog',
     macosStatus: 'Experimental',
     androidStatus: 'Backlog',