+++ apps/fluent-tester/src/FluentTester/TestComponents/TabsExperimental/consts.ts
@@ -1,0 +1,2 @@
+export const HOMEPAGE_EXPERIMENTAL_TABS_BUTTON = 'Homepage_Tabs_Button_Experimental';
+export const TABS_TESTPAGE = 'Tabs_TestPage';