+++ packages/components/Button/src/PrimaryButton/PrimaryButton.settings.win32.ts
@@ -5,10 +5,10 @@ export const settings: IComposeSettings<IButtonType> = [
   {
     tokens: {
       backgroundColor: 'brandBackground',
-      color: 'neutralForegroundInverted',
+      color: 'neutralForegroundOnBrand',
       borderColor: 'brandBackground',
     },
-    // TODO - #728: neutralForegroundInverted is not working for icon color.
+    // TODO - #728: neutralForegroundOnBrand is not working for icon color.
     endIcon: {
       color: '#ffffff',
     },
@@ -26,14 +26,14 @@ export const settings: IComposeSettings<IButtonType> = [
       hovered: {
         tokens: {
           backgroundColor: 'brandBackgroundHover',
-          color: 'neutralForegroundInverted',
+          color: 'neutralForegroundOnBrand',
           borderColor: 'brandBackgroundHover',
         },
       },
       pressed: {
         tokens: {
           backgroundColor: 'brandBackgroundPressed',
-          color: 'neutralForegroundInverted',
+          color: 'neutralForegroundOnBrand',
           borderColor: 'brandBackgroundPressed',
         },
       },
@@ -41,7 +41,7 @@ export const settings: IComposeSettings<IButtonType> = [
         tokens: {
           backgroundColor: 'brandBackgroundHover',
           borderColor: 'strokeFocus2',
-          color: 'neutralForegroundInvertedHover',
+          color: 'neutralForegroundInvertedLinkHover',
         },
       },
     },