+++ packages/components/Tabs/src/TabsItem.tsx
@@ -11,12 +11,7 @@ import { filterViewProps } from '@fluentui-react-native/adapters';
 import { mergeSettings } from '@uifabricshared/foundation-settings';
 import { TabsContext } from './Tabs';
 import { tabsItemName, TabsItemType, TabsItemProps, TabsItemSlotProps, TabsItemRenderData, TabsItemState } from './TabsItem.types';
-import {
-  useAsPressable,
-  useViewCommandFocus,
-  createIconProps,
-  useOnPressWithFocus,
-} from '@fluentui-react-native/interactive-hooks';
+import { useAsPressable, useViewCommandFocus, createIconProps } from '@fluentui-react-native/interactive-hooks';
 
 export const TabsItem = compose<TabsItemType>({
   displayName: tabsItemName,
@@ -39,24 +34,32 @@ export const TabsItem = compose<TabsItemType>({
     // Grabs the context information from Tabs (currently selected TabsItem and client's onTabsClick callback).
     const info = React.useContext(TabsContext);
 
-    /* There's a bug where the user callback is being called multiple times on one click.
-    We check that there is an actual change in selection before forwarding the message to the user callback
-    so that they aren't notified more than once for each click. */
-    const changeSelection = () => {
-      if (itemKey != info.selectedKey) {
+    const [focusState, setFocusState] = React.useState({
+      focused: false,
+    });
+
+    const changeSelection = React.useCallback(() => {
+      componentRef?.current?.focus();
+    }, [componentRef]);
+
+    const changeSelectionWithFocus = React.useCallback(() => {
+      setFocusState({ focused: true });
+      if (!focusState.focused) {
         info.onTabsClick && info.onTabsClick(itemKey);
-        info.getTabId && info.getTabId(itemKey, info.tabsItemKeys.findIndex(x => x == itemKey) + 1);
+        info.getTabId && info.getTabId(itemKey, info.tabsItemKeys.findIndex((x) => x == itemKey) + 1);
         info.updateSelectedTabsItemRef && componentRef && info.updateSelectedTabsItemRef(componentRef);
       }
-    };
+    }, [focusState, setFocusState, componentRef, info, itemKey]);
 
-    // Ensure focus is placed on tabsItem after click.
-    const changeSelectionWithFocus = useOnPressWithFocus(componentRef, changeSelection);
+    const removeFocus = React.useCallback(() => {
+      setFocusState({ focused: false });
+    }, [setFocusState]);
 
     const pressable = useAsPressable({
       ...rest,
-      onPress: changeSelectionWithFocus,
-      onFocus: changeSelection,
+      onPress: changeSelection,
+      onFocus: changeSelectionWithFocus,
+      onBlur: removeFocus,
     });
 
     // Set up state.
@@ -107,10 +110,10 @@ export const TabsItem = compose<TabsItemType>({
         accessibilityLabel: accessibilityLabel,
         accessibilityState: { disabled: userProps.disabled, selected: info.selectedKey === userProps.itemKey },
         accessibilityActions: [{ name: 'Select', label: tabsItemSelectActionLabel }],
-        accessibilityPositionInSet: accessibilityPositionInSet ?? info.tabsItemKeys.findIndex(x => x == itemKey) + 1,
+        accessibilityPositionInSet: accessibilityPositionInSet ?? info.tabsItemKeys.findIndex((x) => x == itemKey) + 1,
         accessibilitySetSize: accessibilitySetSize ?? info.tabsItemKeys.length,
         onAccessibilityAction: onAccessibilityAction,
-        focusable: Platform.OS === 'macos' ? !userProps.disabled : userProps.focusable ?? true,
+        focusable: Platform.select({ default: true, macos: !userProps.disabled }),
       },
       content: { children: headerText + countText, testID: testID },
       icon: createIconProps(icon),