+++ packages/theming/theme-tokens/src/tokens-global.win32.ts
@@ -1,0 +1,3 @@
+import globalTokens from './global-win32/reactnative/tokens-global.json';
+
+export default globalTokens;