+++ apps/web/CHANGELOG.md
@@ -1,8 +1,40 @@
 # Change Log - @fluentui-react-native/tester-web
 
-This log was last generated on Tue, 31 Aug 2021 18:48:34 GMT and should not be manually modified.
+This log was last generated on Fri, 10 Sep 2021 21:30:12 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.1.0
+
+Fri, 10 Sep 2021 21:30:12 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tester to v0.43.0 (30809111+acoates-ms@users.noreply.github.com)
+
+## 0.1.0
+
+Fri, 10 Sep 2021 18:16:59 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tester to v0.42.2 (ruaraki@microsoft.com)
+
+## 0.1.0
+
+Thu, 09 Sep 2021 22:12:19 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tester to v0.42.1 (ruaraki@microsoft.com)
+
+## 0.1.0
+
+Thu, 09 Sep 2021 20:03:01 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tester to v0.42.0 (v.kozlova13@gmail.com)
 
 ## 0.1.0