+++ packages/components/FocusTrapZone/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @fluentui-react-native/focus-trap-zone
 
-This log was last generated on Thu, 26 Aug 2021 21:02:08 GMT and should not be manually modified.
+This log was last generated on Fri, 10 Sep 2021 18:16:59 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.7.44
+
+Fri, 10 Sep 2021 18:16:59 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/interactive-hooks to v0.10.43 (ruaraki@microsoft.com)
+
+## 0.7.43
+
+Thu, 09 Sep 2021 22:12:19 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/interactive-hooks to v0.10.42 (ruaraki@microsoft.com)
+
+## 0.7.42
+
+Thu, 09 Sep 2021 20:03:01 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/interactive-hooks to v0.10.41 (ruaraki@microsoft.com)
 
 ## 0.7.41