+++ packages/framework/framework/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @fluentui-react-native/framework
 
-This log was last generated on Thu, 26 Aug 2021 21:02:08 GMT and should not be manually modified.
+This log was last generated on Fri, 10 Sep 2021 18:16:59 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.5.37
+
+Fri, 10 Sep 2021 18:16:59 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/default-theme to v0.8.3 (ruaraki@microsoft.com)
+
+## 0.5.36
+
+Thu, 09 Sep 2021 22:12:19 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/default-theme to v0.8.2 (ruaraki@microsoft.com)
+
+## 0.5.35
+
+Thu, 09 Sep 2021 20:03:01 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/default-theme to v0.8.1 (ruaraki@microsoft.com)
 
 ## 0.5.34